requirements.txt.txt
streamlit
pandas
numpy
matplotlib
mysql-connector-python
sqlalchemy
ipython
